requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-dotenv>=1.0.0
gunicorn>=21.0.0
PyMySQL>=1.1.0